[PS,MW] Moved tests to test package to exclude them from coverage tests
